requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiohttp>=3.8.0
asyncio>=3.4.3
git+https://github.com/ChrisLolz/gpt4free.git#egg=g4f[all]